search: code monitoring, move monitors to store
With the exception of UpdateCodeMonitors, this PR moves all
db-interactions with cm_monitors from `resolvers.go` to the store.

UpdateCodeMonitors need a little bit of refactoring itself before it can
move down. I will postpone this for now to keep momentum for the
delivery of the POC.

With this PR it will be much easier to create meanigful tests for the
store.