search: add cols num_results, trigger_event
`num_results` keeps track of the number of results found, a number we want
to show in the emails we send out

`trigger_event` links the action events with the trigger event they
originated from. This is necessary to get the actual query we ran and
the number of results based on the action event.