search: prepare store to send emails
To finally send emails, the store needs to provide a few more
methods. I used the oportunity of previous refactorings to add several
tests to the store.

A follow-up PR will contain the changes to the background workers that
will call the methods introduced here.

I also harmonized the contents a bit, moving structs and db-column
definitons to the top of the files, bloating the diff a bit.