repo-updater: Don't delete repos when rate limit exceeded (#16453)
When we perform a repoLookup on sourcegraph.com and discover that we are
not able to access a public repo we delete that repo as we assume that
it is no longer public. Unfortunately, the GitHub rest API returns a 403
code when rate limit has been exhausted.

To fix this we first check for the special case of rate limit error
before falling back check if we have a 403 or 401 error.

This change also recognises when we internally rate limit requests.