migrations: remove unused campaigns tables
In #13001, we migrated campaigns from our old, pre-spec workflow to the
new, spec-based workflow. As an escape hatch for support, we preserved
the old tables in case the migration failed and we needed to retrieve
unmigrated data.

Several versions later, we believe all campaign users have upgraded, and
no issues have been reported with that migration. Therefore, it's time
to burn the shipyards that built the boats we burned in #13001.

Closes #16223.