search: fix paging for monitors
`totalCount` should return the total number of monitors and not the size
of the page we currently return.

I added a test case to make sure we don't regress. However, the tests
are now at a point where the overall runtime (approaching 30s) is
just too long. Hence, I decided to combine all tests that just
query a code monitor and let them share the setup of test data, which
reduced the runtime by 8s.

Tests are now quicker to write and faster to execute. Additionally, new
tests to query monitors come at almost zero cost of execution time, when
compared to the setup time of the test data.

The downside is that they are a bit harder to read.