search: fix totalCount for trigger events
just like #16506 and #16545

The actual change is pretty small, but I had to update the tests to make
sure we catch regressions. The challenge is to create a non trivial
state in the database during tests. Since we are dealing with trigger
events , IE comlpleted jobs with at least 1 search results, I had to
make sure the test data contains at least 1 completed job. This bloats
the diff a bit, but it is for a good cause.