search: delete old job logs for code monitoring
Each time we run a query we log how many seach results it returns.
Currenlty, we delete logs with `num_results = 0` but retain logs with
`num_results>0` forever. This PR introduces a retention policy, which is
set to delete all logs older than 7 days. Once we expose events to
users, we should extend this period to 30 days or more.