Fix racy read on r.namespace
go test -race complained about a concurrent read on r.namespace, which actually can occur when the namespace is read and at the same time the namespace is stored in the sync.Once in computeNamespace.
This fixes it by introducing a preloadedNamespace property, a pattern we commonly use so I think it's fine to do here as well.