search: searchResultsCommon.repos is a map
This is the first in a series of changes to searchResultsCommon to make
it more efficient as well as concurrency safe. We need to make it
concurrency safe such that we can report its intermediate states for
streaming.

This change changes searchResultsCommon.repos from a slice into a
map. In general our use of slices doesn't really make sense, since
whenever we read we do a "dedupSort" which mutates the slice. What we
really want is a set. Then at final read time we can sort.

To support this change we also only set "repos" once, at the time of
repos being resolved. We unnecessarily did it inside of each search
backend. For pagination this works slightly differently, so we still
support updating repos as pagination goes over the repo batches.

We intend to simplify the other searchResultsCommon fields since they
should always be subsets of repos. This is for the purpose of making it
possible to stream progress updates.