search: include sample of skipped repos in message (#16662)
This is to improve the skipped messages. We include a sample of the
repositories skipped in the message. If there is only 1 skipped
repository, we more prominently show its name.

To achieve this change we did some refactoring since there is a lot of
commonality between the skipped handlers involving repos. You can review
this commit by commit.