Introduce store package (#16634)
This introduces a separate store package to our code-base which makes for a nice abstraction IMO. Thorsten and I hacked on this together in our hack session yesterday, and it feels kinda nice having the separation. Also, the root folder of our enterprise code is a little more manageable and file names are less redundant.