Remove httpcli from resolvers and service packages (#16690)
This was never passed in from the outside, and the frontend doesn't have one on disposal it seems. Given that repo-updater does essentially the same thing for the syncer, I think it's fine to get rid of it and just pass-in nil. The external services create their own CLI instance anyways when none is passed in, so this feels like useless boilerplate and like something that can be easily forgotten to pass around either.