searceh: add mutation to trigger actions immediately
Feedback from first testers of code monitoring was that it would be
great if there were the possibility to trigger actions immediately and
not to wait for a commit to happen. This PR adds a mutation that allows
just that. The mutation is (for now) restricted to site admins. The plan
is to add a button to the UI behind a sub feature flag to allow site
admins to troubleshoot or test the configuration.