zoekt: do visibility checks when constructing partial (#16755)
The zoekt code in graphqlbackend would leak potentially private
repository names via the "partial" set we construct. This is since we
directly put names into the partial field before validating if the user
had access to it.

In practice this would not occur since we could never get the repo from
the DB (so the resolver could not be created) and the set of global
repositories only contains public repos.

This PR deduplicates code between structural and normal indexed
search. It also aligns the types closer to what we have in
searchResultsCommon, which makes a future change cleaner.