Add proper preview API (#16535)
This PR introduces new resolvers that, instead of exposing the reconciler plan weirdly on the changeset spec, exposes a proper review API that accounts for detach and close properly as well.

Since we never shipped a release with the previous schema, I think it's fine to just remove the resolvers on changeset spec without any notice.

I know naming is not the best in here and suggestions welcome, but I'd like to land this regardless of naming, before the release. Mainly to get the chance to get some feedback on this API/UI in the next month. We can always follow-up :)