gitserver: fetch gerrit changeset refs (#16767)
We currently hardcode the list of refs we fetch to just the code hosts
we support. The original motivation behind this was to make fetches
faster. However, we have likely reached the point we should fetch
everything. I will leave that for another PR incase it causes
issues. For now lets add in gerrit to the allow list.