Use internal actor in campaigns background jobs
With the switch from `repos.Store` to `db.Repos` in #16753 our
`loadRepo` functions started to fail with "repo not found" errors,
causing the reconciler and the syncer to fail for private repositories.

The reason for that failure was that in `loadRepo` didn't have an actor
set and the repo permissions check returned false.

This change here sets a "global" context to be used by all campaigns
background jobs when talking to the database, which is fine since we
enforce repo permissions in the rewirer when we have the user's actor
(since that happens in the request path).