web: support conditionally rendered inputs in useInputValidation (#16737)
Use a ref callback instead of a mutable ref object so we have a 'notifier' observable that emits when the input reference changes. This is important because the input element can be conditionally rendered, and we can only validate through the Constraint Validation API when we have a reference to an HTMLInputElement.