campaigns: implement licence checks (#16687)
This PR implements the newer licensing behaviour discussed at sprint planning: campaigns may be used without a licence for campaigns with no more than five changesets, but attempting to create a campaign spec with more than that results in an error. We don't enforce this for retired plans, since those licenses include access to ALL features currently. This implementation does not respect the EnforceLicensing flag since we want to do this earlier.