search: refactor alertOnSearchLimit
Currently, alertOnSearchLimits mutates resultTypes and acts as an
implicit `if`. With this change we move the check for search limits inside
the diff/commit searchers. The searchers return a `multiErr` which we
convert to an alert afterwards (similar to how we do it for structural
search).

This is nice because we push the alert handling to the end of `doResults`.
In the future we can extract this logic out of `doResults` and let
it simply return an error.

I split the existing test case in two test cases, one checking if the
correct error is returend, and one checking if the error is converted
into the correct alert.