Merge unsynced and publication state together (#16860)
Unsynced can actually be derived from publication state, so why have two different ways for branch and imported changesets. This PR merges the two into a unified solution, making the publication state the single source of truth. Also, this makes more sense with the default value we have for the column in the database.