Don't update changeset diffstat on changeset spec change
The diff is not yet updated at this point and hence it caused confusion in the UI where the diffstat was out of sync with the actual diff.
Since `SetDerivedState` is called after all code host operations and that one does `SetDiffStat` anyways, this just defers updating the diff stat of the changeset to after the commits have been pushed.