Fix indenting of "Clone the deploy-sourcegraph" repository bullet item
Adding spaces to the fenced code block for GCP kubectl config.  Doing this so the "Clone the deploy-sourcegraph repository" step which follows is a bullet in the main list rather than a sub-bullet of "Make sure you have configured kubectl to access your cluster".  
Reference: https://gist.github.com/clintel/1155906