Link in sqlite3 when building libsqlite3-pcre (#17024)
Quinn Keast ran into an issue where compiling this would fail because of
missing sqlite3 symbols. This fixes that error by linking in `sqlite3`,
which I think was/is implicit if sqlite3 is installed globally, but that
might not be the case on a multi-user homebrew setup on macOS Big Sur
(which is what Quinn has).