search: Fix bad revision defaults to main branch
When a revision is specified that does not exist, the resolved
repository contains an empty slice of RevisionSpecifier. This is
different behavior than when no revision is specified, in which case
the resolved repo will contain a single RevisionSpecifier set to the
zero value.

When `type:commit`, the `git diff` arguments are constructed by
filtering down to the revision list. In the case there are no revisions
in the revision list, there are no filtering args added, which results
in the same behavior as the default branch.

This commit skips any repos that have an empty revision list.