Fix match for commit and diff merged into one
When `(type:commit or type:diff)`, if the query matched the same commit
for both a diff and a commit message, only one result would be returned
because we merge `or` results by URL and the URL is the same for the
diff and the commit that the diff is for.

This commit merges diffs and commits separately so that if a query
matches a commit's message and a commit's diff, both will be returned.