search: Fix pageinfo returned for code monitors
The PageInfo returned by the GraphQL resolver for monitors was always
indicating that there were more pages after the first query because it
simply checked if the last result was empty.

This change adds logic to query the number of remaining code monitors so
that HasNextPage is set only if asking for another page would result in
more code monitors.