streaming search: log events (#17138)
Log all events already logged by non-streaming search when using streaming search. Since streaming search was a rewrite of the search results UI, a lot of events that were logged by the old UI were not yet being logged in the new one.

Also added logging for when the user clicks the "Search again" button in the skipped results popover.