search: zoekt streams per replica
This switches the interface between Zoekt and Sourcegraph from using the
batch interface to the streaming interface. Now if you use streaming
search you will receive search results as each replica completes.

This is an intermediate state. We will clean up the several layers of
channels in between that just shuffle data around to a use an unified
event.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>