Only set DisplayName and AvatarURL if empty. (#17257)
If a user signed in via multiple code hosts with different display names and avatars, their Sourcegraph experience would flip-flop between them as they logged in via the different hosts. This changes that behavior so that updates will only be made to the underlying data when it's absent. Otherwise the first-encountered values will be used.

fixes https://github.com/sourcegraph/sourcegraph/issues/17086