search: use an enum for repo status fields
This more directly represents the different states a repository can be
in. This change was motivated by making it easier to communicate
progress updates. Additionally we avoid the need for our dedupSort call
when creating a list of resolvers.

We have an assumption where searchResultsCommon.repos is a superset of
the status map we create. This assumption only needs to hold when
searching is done.

For search we set searchResultsCommon.repos before any backends can
report search results. For paginated search we have included a fix which
sets repos before any searches. This is likely a fix for a regression in
paginated search.