streaming search: run AND expressions in batch mode (#17276)
streaming search was broken for expressions because we streamed back results of individual terms. While this is desirable for OR expressions, it breaks AND expressions because we are only interested in the intersection of their term results. With this change, we fall back to batch mode for AND expressions, streaming back results only after we calculated the intersection of all term results.

In the future, we might want to handle this more efficiently by calculating the intersection on the fly.