search: prepare extraction of SearchResultsCommon from graphqlbackend (#17315)
We want to extract SearchResultsCommon out of graphqlbackend. This
exports everything to allow that. Additionally it disentangles the
graphql resolvers from the struct.

Additionally we remove an effectively unused `resultsCount` field. This
field was never exposed to a user. Additionally the only read path was
in an or condition with `limitHit`. However, it could only be true if
`limitHit` is true, so was a codepath that would never return true.