search: Add zoekt querying to searcher
Adds the ability of searcher to query zoekt directly for the file list
and contents that it should perform structural search on.

To activate that code path, a temporary special comby rule
(`rule:'where "zoekt" == "zoekt"'`) needs to be added to the query.

There are no frontend changes included here, so the frontend will still
query zoekt to get the list of files, but searcher will ignore that file
list if the special rule is set.