search: streaming interface for zoekt (#17245)
This introduces a streaming interface for Zoekt called
`StreamSearcher`. This adds an event based equivalent of zoekt's batch
method `Search`. We then update HorizontalSearcher and MeteredSearcher
to be stream based. Additionally we add an adapter which takes the batch
interface and "upgrades" it to streaming. This allows us to update most
places to instead take a `StreamSearcher`. The next PR will actually use
this streaming searcher, which will allow us to stream results as we
receive them from a zoekt backend to the client.

We are embedding `zoekt.Searcher` into `StreamSearcher`. It is weird to
keep supporting the batch interface when we shouldn't be using it going
forward. We intend to disentangle this embedding in a few commit times
once the zoekt batch interface is no longer used.

Test plan: Unit tests as well as manual testing in my dev environment. I
ran a few searches to and inspected the traces as well.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>