Switch 'src validate' to use yaml.v3 (#431)
This fixes #429 by switching `src validate` to use `yaml.v3`.

When trying to confirm that it still works, I found out that parsing the
validate input as YAML didn't work for me, even without the change.

Both `yaml.v3` and `yaml.v2` assume that the fields in the YAML are in
`snake_case` and need tags added to the struct fields to make it work
with `camelCase`.

The `github.com/ghodss/yaml` works with `snake_case` and also supports
the unmarshalling into a `*json.RawMessage`, which `yaml.v2` and
`yaml.v3` don't support.

But since the goal was to only use a single YAML library I switched to
`yaml.v3` and changed the code to work with it. I tested this locally
and checked that everything's unmarshalled correctly.