Remove container restart alerts (#17422)
I tried to find alternatives, but I can't find any that fit both compose and Kubernetes. We might have to make do with some alerts only being available on advanced systems like Kubernetes, definitions for both, or a combination of those things.
Any alert that does not know about deployments and/or desired replica count was not accurate enough to avoid alerts during deployments. While the metric itself is accurate, a container has not reported for X amount of time, we cant identify if that is an OOM/panic or a container has been gracefully stopped for deployment as both cause the same effect.

cAdvisor also keeps container data for 5m after a container is removed, which is why we see the number increase for containers removed by a deployment up to almost 300.

This also renames the metric/panel to Containers missing.

Co-authored-by: Robert Lin <robert@bobheadxi.dev>