dev: run codeinsights-db (TimescaleDB) as part of dev server
We plan to deploy TimescaleDB as a standalone container in our deployments
as the database backend for Code Insights, see: https://github.com/sourcegraph/sourcegraph/issues/17218

This runs the re-tagged `sourcegraph/codeinsights-db` TimescaleDB image
in our dev environments. This is different from how we run Postgres, where
we require devs to install and run it themselves. I think that running it
in Docker is nicer for most devs, and we can always add an option to disable
it later if anyone is interested in installing/running it on their machine
seperately.

Note that one cannot generally connect it to the same Postgres DB, because
TimescaleDB is often version-incompatible with the Postgres version we run
in dev environments (9.6 or latest 13, TimescaleDB only supports 11/12)

Part of #17226

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>