search: enable symbol search by default (#17425)
If you don't specify a type we now include symbol results by
default. This was a feature we removed a while ago due to performance
problems. However, symbols have been indexed for many months so this
feature should be safe to enable again.

This has a low chance of causing production issues, so not hiding behind
a feature flag. Will monitor the performance impact in production and
revert if there are surprises.