search: use SearchEvent instead of slices of results
This updates all uses of our search stream channel to instead use a new
type SearchStream. This required updating both the text and commit
searches which support streaming. For text we adapt it to continue using
the old stream interface to update in a later commit. For commit we
directly use the new stream.

This commit is in preperation for streaming progress updates.