search: remove backend filepath dedupping (#17466)
This was done since when we include both symbol and text search results
we want them to appear under the same file. However, we currently do not
include symbol results by default. Additionally the logic for merging
needs to live in the client going forward since we are moving to a
streaming interface.

Our frontend merging correctly works when we return a results for both a
filename and the contents of the file.

Additionally we want to avoid unbounded state while streaming. This is
one such state since it is the size of all file path results streamed
back.