insights: add GraphQL backend scaffolding (#17497)
* insights: add GraphQL backend scaffolding

This adds the needed GraphQL backend scaffolding for code insights. The actual
schema here is not important and will definitely change.

I've based this on what we do for Code Intel, Campaigns, and Code Monitoring.
It seemed a large enough change (without actually implementing any GraphQL APIs)
to warrant sending separately. This way, PRs that implement actual APIs for
Code Insights can be reviewed on solely that, rather than all the scaffolding that
is required.

What this change does:

* Registers stub GraphQL resolvers for non-enterprise Sourcegraph versions.
* Injects GraphQL resolvers for enterprise Sourcegraph versions.
* Prepares to (but does not yet) run database migrations for code insights.

Helps #17221

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>