search: aggregator is stream based (#17470)
We update the aggregator to be stream based. This is mostly a design
change, but functionally we now ensure that every error and stats is
reported on the stream. This will allow us to accurately report progress
from the HTTP layer down the the client. Additionally, we are now much
closer to making the graphqlbackend for search purely stream based.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>