search: add an explicit searchFilesInReposBatch function (#17504)
We want to make the normal interface for searchFilesInRepos to be
streaming. So it no longer makes sense to have return arguments. So we
introduce a batch interface which doesn't take a stream and does the
collection for us. I made the collection generic since I think we will
likely use this pattern more than once.

Additionally this makes our hacks for structural search clearer. I add
back the special logic around unsetting IsLimitHit.