search: stream stats from text backends
This was the final hurdle preventing real streaming of progress
updates. Now we need to update the HTTP layer to stream out the progress
events. We remove the aggregation of stats inside of
doSearchFilesInRepos and instead stream the stats as we receive them.

We also went along and cleaned up the function a bit. Instead of
duplicate handling around limits and errors at each call site, we move
the logic into send. Additionally, we remove the need to hold the mutex
before sending down the channel. We now only hold the mutex when
accessing shared fields.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>