search: cleanup resources on early return with global search
We did not wait for global searches to finish when we did an early
return. This lead to a panic since the stream channel would be closed
followed by global search writing to it.

I don't think this is a long term solution. Rather we need to move the
interactions between stream and search backends into a clearer
relationship to prevent these sort of bugs.

We also remove the unnecessary cancel context in global search. We are
already protected by the cancel context in doResults. We rely on this
context to correctly cleanup for any exit.