search: stream progress to webapp (#17531)
We add Stats aggregation to the top level streaming layer. This allows
us to stream out progress updates as they happen.

This can be made way more efficient. There are also some changes that
need to be made to the API/webapp. However, this is now functional and
only affects people who enable streaming. So I would like to get this
out so we can play with it in production.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>