search: only return limit results from each zoekt (#17553)
Previously we would ask for 2 * limit results from Zoekt. However, we
would just throw away every result over limit. The documented reason for
asking for more was for more accurate limit hit per repo statistics. To
avoid asking for a lot more in the case of a large limit, we instead
always ask for 2000 more than the limit.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>