Make changesetStats use the same logic as changeset.state (#17543)
Before, the logic was a little different so in the stats bar you would see 10 failed, but when filtering by state failed, you might get 0 results, which gives the impression some results might be hidden from you. This fixes it by using the same conditions for the different statuses. Also, I've extended the type by the missing states, just so the SQL query is complete and easier to compare. I don't use it _yet_.