enterprise/internal/insights: add basic store package + testing infrastructure
This adds the basic store package implementation (just the groundowork, no
actual implementation) based on what we do in campaigns, code monitoring, etc.

I am sending this as a separate PR before implementing the store, as it is
a non-trivial amount of setup code which is nice to track separately for
purposes of documentation and for seeing how we can improve adding backends
like this, campaigns, code monitoring, etc. in the future.

Helps #17218

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>