Add tests for the changeset rewirer
Since we split up packages, there's no reported coverage on this piece of code anymore, and that was actually on purpose, to force me to do this at some point.
This introduces DB state independent tests for the rewirer, so they are easy to write and maintain.
Ultimately the goal should become to reduce the tests for service.ApplyCampaign to a bare minimum, and have each component tested individually.