code insights: add dev option to disable code insights backend entirely
It seems that for some reason our PostgresDSN code picks up `PGDATABASE`
when in reality it should only be picking up `CODEINSIGHTS_PGDATABASE` for Garo
as seen in [this Slack thread](https://sourcegraph.slack.com/archives/C07KZF47K/p1611801989103400).

Probably the underlying pgx driver does this, and we just didn't notice it
with codeintel before because it uses the same DB whereas codeinsights does
not.

I will fix that, but this option immediately unblocks Garo and others who
might run into this - and it's a useful escape hatch anyway.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>