dev: give codeinsights-db more time to start
The 10s timeout is pretty aggressive when a bunch of services are starting
and leads to codeinsights-db not having enough time to start before the
frontend gives up, so raise it to 30s in dev environments only.

If we had a better way to describe service dependencies with goreman,
we could avoid this - but this seems harmless.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>